feat: added the ability to add new auctions to the auction house through the addauctionform. The user can now see their new auction. Still need to work on getting rid of the item from the user's inventory and make sure it's fully updated correctly when placing a new auction.

+++ src/main/java/com/raistmere/fantasyauctionhouse/repos/InventoryItemRepository.java
@@ -3,5 +3,5 @@ package com.raistmere.fantasyauctionhouse.repos;
 import com.raistmere.fantasyauctionhouse.domains.InventoryItem;
 import org.springframework.data.repository.CrudRepository;
 
-public interface InventoryItemRepository extends CrudRepository<InventoryItem, Integer> {
+public interface InventoryItemRepository extends CrudRepository<InventoryItem, Long> {
 }
